requirements.txt
beautifulsoup4==4.12.3
requests~=2.32.4
pillow==10.4.0
imagehash==4.3.1
pyyaml==6.0.2
numpy==2.0.1
scikit-image==0.24.0
Flask==3.0.3
dmca_monitor~=0.1.0
APScheduler>=3.10